requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Configuration
python-dotenv>=1.0.0

# RSS and Web Scraping
feedparser>=6.0.0
trafilatura>=1.6.0
requests>=2.31.0

# Text Processing and Analysis
sumy>=0.11.0
vaderSentiment>=3.3.0
nltk>=3.8

# Data Validation
pydantic>=2.0.0

# Utilities
python-dateutil>=2.8.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.0.0
pytest-cov>=4.0.0
